done: flashWithdrawal and setters
